Fix advanced search for section language
In the advanced search, when I filter for a language, the search result
not only lists sections in that language (which I want) and projects
containing a section in that language (which I also want), but also
sections from multilingual projects that are in other languages (which
I'm not interested in). This patch removes those sections in other languages
from the search results.

Change-Id: Ie02c80fe49a937ec842a6c9cff0ccc73843c64a7