Fix advanced search for section language
Given the following scenario:

    - a multilingual project
      - an Arabic section
      - a Chinese section

When I browse for "language Arabic" in the catalog, the URL is:

    /search/get_results
      ?primary_key=4
      &search_category=language
      &sub_category=language

and the search result is:

    - a multilingual project
    - an Arabic section

But when I do an "Advanced search" for "language Arabic", the URL is:

    /advanced_search
      ?recorded_language=4
      &search_form=advanced

and the search result is:

    - a multilingual project
    - an Arabic section
    - a Chinese section

This is unexpected.

This commit makes the two search results equal by only showing sections
of the selected language. Sections in other languages that happen to be
part of a multilingual project containing the selected language are no
longer shown.