Fix advanced search for section language
In the advanced search, when I filter for a language, the search result
not only lists sections in that language (which I want) and projects
containing a section in that language (which I also want), but also
sections from multilingual projects that are in other languages (which
I'm not interested in).

Given the following scenario:

    - a multilingual project
      - an Arabic section
      - a Chinese section

When I browse for "language Arabic" in the catalog, the URL is:

    /search/get_results
      ?primary_key=4
      &search_category=language
      &sub_category=language

and the search result is:

    - a multilingual project
    - an Arabic section

But when I do an "Advanced search" for "language Arabic", the URL is:

    /advanced_search
      ?recorded_language=4
      &search_form=advanced

and the search result is:

    - a multilingual project
    - an Arabic section
    - a Chinese section

This is unexpected.

This commit makes the two search results equal by only showing sections
of the selected language. Sections in other languages that happen to be
part of a multilingual project containing the selected language are no
longer shown.

Change-Id: Ie02c80fe49a937ec842a6c9cff0ccc73843c64a7