Remove third_party directory
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

In preparation for the renaming of classes to Ucfirst case (step 2 of
the upgrade), remove the third_party folder. It appears to contain
unit tests, which in practice have no use anymore.

[1] https://codeigniter.com/userguide3/installation/upgrade_300.html