Rename class files to Ucfirst case
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

Step 2 of the upgrade is to rename class files to Ucfirst case. This
commit does that, and updates their corresponding `require`s.

[1] https://codeigniter.com/userguide3/installation/upgrade_300.html