Move error templates to views/
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

Step 8 is to move error tempates to views/. This commit does that.

In addition, CodeIgniter 3 added a new error_exception.php template as
well as cli versions of each template. Since we did not have any
customizations to the templates anyway this commit just copies all the
new reference templates from CodeIgniter 3.

[1] https://codeigniter.com/userguide3/installation/upgrade_300.html