Autoload security helper
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

Step 13. Since the xss_clean form validation rule is sprinkled all
over the code an we're not going to undertake the massive effort to
clean that up, just autoload the security helper.

[1] https://codeigniter.com/userguide3/installation/upgrade_300.html