Re-add custom idiom to application/
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

The `alpha_dash_space` idiom was in system/ for some reason, and got
clobbered in the first step of the upgrade. This commit re-adds it in
application/.

[1] https://codeigniter.com/userguide3/installation/upgrade_300.html