Add Copy to Clipboard button to Project Launch Template result page.
This seems like a good feature to add as it can be a little tedious to
manually select and copy all the text and should result in less errors.

While adding the button I also did some general cleanup of the page layout.
I found it annoying how the text area had a larger height than the page
on my laptop resulting in two scroll bars (one for the text area and one
for the whole page). So I made the text area height dynamically size to
the available area (with a minimum size if the window is really small). I
also found the horizontal rules before and after the text area and the
footer seemed to be mostly a waste of space resulting in less of the
results being visible so I removed them.