Fix image link href in browse catalog title blade.
The image link href was hard coded to "#". Now it links to the
corresponding title.

The Title URL was already being built up in two different spots. This is
now done once and stored as $url and used for the image and the other
spots. The image source logic was also refactored into the same set of
if statements to make the code clearer.