Improve formatting of filenames for project launch template.
Issue #25 asked for parenthesis to be removed from filenames. The current
code had a short black list of special characters to remove. However it
seems better to instead have more of a whitelist approach as there are many
more special characters that should also be removed.

This commit uses the ICU transliterator (from php-intl package) to convert
all characters to ASCII which removes all special symbols and accents and
then removes anything (including spaces) that is not a-z or 0-9. It will
even convert things like 汉语" to "hanyu" ("han yu" before spaces are
removed).

Previously the limited black list was also only applied to the title and
not the author portion of the filename. It is now applied to both parts.

This commits also renames the array keys titleid to title_filename and
authorlc to author_filename which I think are more clear.