Fix copyright notice when no translator and other code cleanups.
When the translator list was empty the code was executing:
 $data['notice'][0] = '';

This erased the copyright notice from the first author.

The code was explicitly initializing $data['author'] and
$data['translator'] as arrays. I added similar initialization for the other
entries such as authorall and translatorall. This is technically not needed
but this makes the code more consistent and it is considered good practice
to explicitly create arrays. I also removed adding blank entries to these
arrays when the author or translator lists are empty.