Use get() instead of get_post()
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

In CI3, it looks as though get_post(param) needs a parameter name to
return anything. Put differently, our current usage of get_post(NULL)
returns an empty array. Change this to get(null), which works like we
expect.

[1] https://codeigniter.com/userguide3/installation/upgrade_300.html