Several project launch field validation updates and other tweaks.
Make the number of sections mandatory and enforce a numeric value.

The link to text should be optional if the project is a compilation. When
compilation is selected the link to text field is now hidden. The is a
compilation question was moved before the link to text field which is more
logical when link to text depends on it.

The Other language field was only shown if the language was "Other".
However at some point it looks like the language was renamed in the
database to "Other/Uncoded" so this did not work anymore. Now the field
is shown if the language starts with "Other".

I also noticed that if you press the Clear form button, refresh the page,
or go back from the results page the fields such as Target complete date,
Other language and Link to text are not hidden/shown based on the current
field values. A custom handler was added for the Clear form button and the
code updated to reset the visible status of these controls on page load.

While working on this I also noticed that the project launch template is
the only form to use a javascript form validator instead of the CodeIgniter
validator functions. (It actually contains both. The javasript one runs
first. If the CodeIgniter one fails some of the fields are lost such as
author which is probably why the javascript one got added?) I also noticed
the soloist name field had an "alpha_num" rule. However this seems to be
a typo of "alpha_numeric". However this field should not be restricted
to alpha numeric anyway so I removed this rule.