Fix missing dramatic work project type option when using Dutch language.
The create_array_from_lang function used to generate the list of strings
for the project type drop down (and some other drop downs) was written in
a very confusing way for no apparent reason. It expected the language
array to be passed in a flipped state, filtered the array and then flipped
the results before further processing.

The original flipping of the language array causes the Dutch string for
dramatic work to be lost as there happened to be two strings with the same
translated value. The only way to prevent this (without changing the
translations) is to avoid flipping the array.

I re-wrote most of the create_array_from_lang fuction with one simple
foreach loop that works on the (un-flipped) language array. All spots that
called this function had to be updated accordingly. This seems far easier
to understand, is probably faster and doesn't break on the current Dutch
strings.