Allow line breaks in summary on project launch form.
The admin page for editing a project allows line breaks so what not the
project launch page?

Previously the project launch form had a javascript function to block all
use of the Enter key (probably to block the form being submitted on
Enter?). This also blocked the Enter key in the summary field. Instead I
now changed the submit button to be type="button" and added a handler so
that this button still submits the form. This is more similar to how the
admin page was implemented.