Upgarde to Ion Auth 3
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

To work with CI3, we need to upgrade Ion Auth from 2 to 3. This mostly
follows [2], with the exception of the controllers/Auth.php file,
which is "too custom" to be a simple drop-in update. For the DB
migration, two users have the admin@librivox.org email address. Ion 3
enforces unique email addresses, so we chose to remove the 'librivox'
user and keep 'admin'.

[1] https://codeigniter.com/userguide3/installation/upgrade_300.html
[2] https://github.com/benedmunds/CodeIgniter-Ion-Auth/blob/3/UPGRADING.md