Re-order project_launch_template_lang strings.
A script was used to re-oreder the project_launch_template_lang files
to match the order they are used in the code. This tends to group
things together more logically and makes translation a litte easier.

The results of this were also exported to a Google docs spreadsheet for
the admins to update the translations.