Add more strict filename test in validator (including check for "_128kb").
The admins had requested that a check for "_128kb" be added to the
validator. Simply adding "_128kb" before the ".mp3" in the existing
regular expressions would cause many existing filenames to be rejected.
The existing regular expressions actually let almost anything close to a
valid filename through. For example the "_[\d-]" portion of the expression
which seems like it was intended to match the section number could also
match the "128kb". So I basically started over on the regular expressions.

There does not seem to be a good definition of exactly what the filename
patterns should be. For poetry the vast majority of filenames are
"title_author_initials_128kb.mp3". For other non compilation projects the
vast majority of filenames are "title_##_author_128kb.mp3". For
compilations everything seems to be all over the place. So for these first
two cases it now looks for those strict patterns. For compilations it
only checks that it ends in "_128kb.mp3".

To improve this further the admins probably need to come up with a more
detailed set of rules.