Fix validator when there are no files.
If there were no files in the directory and the validator was run the user
would get a loading dialog indefinitely. The load_project_map function
was testing for an empty file map and returning false in this case. I have
now removed this test. The code works fine when $map is empty and will
correctly report to the user the missing files in the completeness tests.

While testing this I also noticed the reported file numbers were wrong in
the error messages for missing files which is also now fixed.