Move number of sections question after preface question.
The number of sections should include the preface. The "Number of sections"
text now indicates this so it makes more sense to have the preface question
first.

The alignment of all three input elements in this section of the form were
also made to be the same which looks neater. The number of sections box
needed to be moved over anyway for the new longer text.