Rework and consolidate author display formatting in PHP code.
Similar to a recent commit to consolidate all author formatting code
that exists in the JavaScript code this commit tries to do something
similar in the PHP code.

There were many spots in the PHP code that formatted authors for
display with many inconsistencies. Some spots converted last names to
uppercase some did not. (Uppercase last name may have been desirable
in some cases but the application of this was inconsistent. It also
looked weird when the last name is actually something like "Various"
or "Unknown".) Some spots would show birth/death years as "( - )" if
they were not set in the database some spots left it out. Some spots
used the dob-dod CSS style for years some did not. Some spots showed
"Various", some "Various Authors" and some used "and others" when the
number of authors exceeded some threshold. And more...

After consulting with some of the admins the new code uses the
following rules:
1. Do not ever convert last names to uppercase.
2. Always use the CSS style for birth/death years.
3. For multiple authors show up to two (separated by "and"). If more
   than two only show two (without "and") and add "et al.".

All spots (unless I missed some) now use the format_author function
(for contexts where is is always one author) or format_authors (for
contexts where there is a list of one or more authors). The
format_author function is just a one liner that can calls
format_authors. These functions take a flags parameters to control
things like whether to includes the years (if present), if the output
should include links, max number of authors in a list, etc.