Use set_value() correctly and add missing validation rule
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

If the field name is an array, it must end with a '[]' when passing it
to set_value(). Looks like this only started being a thing/enforced in
CI3? Start doing this correctly.

Also add a missing link_to_trans validation rule.

[1] https://codeigniter.com/userguide3/installation/upgrade_300.html