Validator bitrate test improvements.
When displaying the detected bitrate the value from getid3 was being
divided by 1000 and always rounded up using ceil function. For some
128k CBR MP3 files encoded by newer versions of LAME getid3 was
(incorrectly) reporting bitrates just slightly above 128000 such as
128004. This was then being displayed as 129. This has now been
updated to use the round function instead so 128004 would be rounded
to 128. But in some ways this is even more confusing to the user as
it can show 128 but still fail the validation which is looking for
exactly 128000. So if any rounding actually occured a "~" is
prepended to the value. So 128004 is now displayed as "~128" where as
128000 would be displayed as "128".

It is possible, although unlikely, that a VBR file could be reported
as exactly 128000 bitrate. This would have passed the validator test
even though the LibriVox standard is to use CBR. The bitrate test now
checks that getid3 reports the file as CBR.

With the above changes it would actually be possible for the
validator to show the file as 128 but still fail the bitrate test if
the file is VBR. To avoid this possible confusion the validator now
just shows VBR if the file is VBR and only shows a number when it is
CBR.

The validation is still checking for exactly 128000. When the file is
a proper CBR file at a standard bitrate getid3 should return the
exact integer bitrate so this test should be ok.

However there is a bug in the verion of getid3 currenly used that can
incorrectly report some files encoded by LAME as VBR when they are in
fact CBR and also reports the wrong bitrate. So even with the above
changes the validation can still fail when it should not.