Fix uploader link in project launch templates.
The uploader link should be /login/uploader not just /uploader otherwise
you get an error page if you are not already logged in.

The proposed prototypes for the updated project launch templates had the
the wrong link for the uploader which I included in the recent update. I
also ended up with the link in the translation files instead of the main
PHP code. I have removed the link from the translation files and the
correct link is now included by the PHP code more similar to how it was
before.