Adjust spacing of authors/translators section on book page.
Previously the translators were displayed directly after the authors and
there was no space between the authors/translators and the book
description.

The CSS was updated to add 15px margin to the bottom of the author section
similar to how it was done for the description section.

I also fixed a small white space issue with the group links and changed
the description element to a <p> instead of <div> to be more consistent
with the author and group pages.