Use isset to check for translator existence
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

The previous conditional was just `if ($translators[0])`, which
obviously doesn't work is $translators is empty (meaning, there are no
translators and the book is in its original language). We're not sure
how it ever worked, but this patch fixes it with an `isset`.

[1] https://codeigniter.com/userguide3/installation/upgrade_300.html