Fix Project Launch submission with Chinese language.
The language id for Chinese Simplified is zh_cn. But the form validation
was checking for "alpha" only (which fails with the underscore). This
check should be removed. As this id is set by a drop down it does not need
very strict validation anyway, only enough to avoid security issues for
someone who tries to abuse the APIs directly.

I accidentally missed this change in the commit which added the Chinese
language. This change was present on my test server (along with an
incomplete Russian translation and other changes). When I created a new
branch to add just Chinese I missed this change.