Relax validator bitrate test.
getID3 does not always return exactly 128000 even when every frame in the
MP3 file is 128k so lets only check to the nearest 1000.

This commit also gets rid of the "~" that was displayed in front of the
divided by 1000 bitrate when it was not an exact multiple of 1000. Since
getID3 is not always returning 100% accurate values showing this "~" is
meaningless and confusing.