Remove colons from uploader/index.php and move them to translation files.
The recent Chinese translations already had the colons in the translation
files which results in double colons. There was one example of this in
the Portuguse file as well. It is better to have them in the translation
files as different languages have different requirements. For example
Chinese uses a different character code and in French there should be a
space before the colon.