Fix errors in home page search
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

Use `isset` to check for variables being defined and not null.
Previously, this was a simple `if $variable`. We're not sure why
`isset` was not needed before.

Resolves #106