Fix errors in authors page
(This commit is part of the series to upgrade to CodeIgniter 3. The
series follows the upgrade guide [1]. Some steps are skipped as they
are not applicable to our environment).

Much like the previous commit, use `isset` to check for variables
being defined and not null.  Previously, this was a simple `if
$variable`. We're not sure why `isset` was not needed before.

Resolves #107