Fix PHP error on Project not found page.
This commit moves setting of search_category and primary_key closer to the
beginning of the index() member function of the Page class. These should be
set before the call to _render. Previously when certain errors happened
such as the project not being found these were left unset triggering a PHP
error in header.php.

Resolves #111.