Validator: chmod files regardless of how they were imported
Previously, we were only chmod'ing files if they were copied locally
from uploads. This meant that files downloaded from a remote server
were left with mode 0644, and the librivox group (of which the admins
with SSH access are members) did not have write access. This patch
changes this to do the chmod'ing regardless.