concat_date function improvements
The concat_date function would return somewhat weird results if any of
the year, month or day parameters were zero. For example, in the generated
project templates the date would be shown as "0-12-00" if the user did
not select a date. The code in Administor_projects.php avoided this by
explicitly checking if any of these components were zero. This check
has now been moved to the concat_date function itself so that the logic
does not need to be repeated in multiple spots. (This function is currently
only used in two spots.)

The updated concat_date function now also makes more use of the standard
date and mktime functions instead of more simple string manipulation. This
increases the chance of getting a valid date as a result with more
consistent formatting. For example, 2021-02-29 (not a valid date) will now
get converted to 2021-03-01. See mktime documentation for more info.

The concat_date function now also accepts an optional parameter (defaults
to 'Y-m-d') to allow other date formats to be used.