Use entered date in poetry project launch template output.
Previously the poetry templates would always show "Sunday, DATE" (or a
translated equivalent) as a placeholder that then needed to be manually
updated.

Note that the generated date text in the template output is in English
even if another language is selected. It could be a lot of work to generate
these strings properly in all the languages and currently the poetry
templates are only being used for English. Even if these are used for
non English the results are no worse than before as they already needed
to be manually updated to replace the fixed placeholder.

This fixes #124.