Remove custom alpha_dash_space validation rule.
This is a bit of a strange rule. Normally you would uses dashes instead
of spaces when spaces are not allowed. So why have a rule that allows only
alpha, dash and space?

After the previous commit there was only one instance of this rule still
being used for the "proof_level" field in the project launch form. However
spaces are not actually needed here so this was changed to use "alpha_dash"
instead similar to some of the other drop down box fields.

With the above mentioned changed we can remove the "alpha_dash_space" rule
as it is no longer used. And being a bit of a strange rule it is unlikely
to ever be needed in the future either.

Since the "alpha_dash_space" rule was the only logic in
MY_Form_validation.php the entire file has now been removed.