Many updates to how fields are shown/enabled in the user profile screen.
This started out as an attempt to implement the UI side of issue #83 which
requested that only an MC (or admin) should be able to change the Display Name,
Website or Max Project Count.

At first it seemed like it would be easy to simply disable these fields based on
your permissions. However the same form is used for adding users as for editing
a user profile and these fields should not be disabled for the add case.

While working on this I also came across some other bugs and inconsistencies in
these screens. Some examples:
1. Normally only admin can change the username. But if you first did Add New or
   Add Reader and then went to an edit profile screen then anyone could change
   the username (as the username field was made editable for the add and not
   disabled again later).
2. If an error was displayed after pressing Save and then you closed the screen
   then the next time you went into the screen the now irrelevant error message
   was still there.

Configuring the form for edit profile, add user or add reader were previously
all handled by mostly separate code. I combined this all into one function named
show_form which seemed to make things simpler overall. There is now only one
spot where any field is shown/enabled such that you should always get consistent
results (you cannot forget to re-enable a field for one case when it was
disabled in another function).

Fixing all of these things involved changing most of the lines in
profile_modal.js. Since most of the file was changing anyway I went ahead and
updated the white space on all lines to be consistent throughout the whole file.

Fixes #83 (along with an earlier commit).