Disable roles you cannot change instead of hiding them.
Previously you could not see roles you cannot change. But it is nice to
see the current values even if you cannot change them. This also makes
it more similar to how other fields on the profile screen are disabled
when you cannot change them.

Note this only applies when any of the roles are displayed. If you
could not see any of the roles before you still cannot. But if you
could see at least some of them you can now see all of them.

This was quite useful when testing the previous commit which
fixed some bugs with how roles were being set.

This commit also changes the terminology used on the screen by
removing the "Add" text. When the check boxes were unchecked you
are actually removing the role which was not 100% clear when the
"Add" text was present. Not adding a role is not the same as removing
it.