Several fixes and improvements to update_profile API.
This API did not do a very thorough job of only allowing fields to be changed
that you actually have permissions to change. For example previously an MC
could change anyone's password (even the administrators).

Issue #83 also had some new requirements to only allow MCs to change Display
Name, Website and Max Projects. The API now also implements this restriction.

The form validation was also partially incomplete. And the results of
functions like trim and xss_clean in the validation were not actually being
used. The form validation should really only be validating fields that you are
allowed to change. If the fields in the HTML form are disabled (because you are
not allowed to change them) they are missing from the POST and will cause
validation failures.