Only send certain whitelisted fields in get_profile API.
Previously it was sending every field from the users database record
including things like the password hash. The security risk was somewhat
minimal as only Admins and MCs were allowed to use this API for arbitrary
users.

Some other small cleanups were also done in the get_profile function. For
example we only call has_permission once and cache the results.

The elements function used to implement the whitelist also needs the data
in an array and not an object. So small changes were needed to accommodate
this throughout the function.