Several fixes to update_profile/add_profile API for group handling.
Similar to how the previous commit fixed several bugs with permissions on the
basic user profile fields this commit improves the APIs for setting the group
permissions.

For example anyone could previously change the Readers or Proof Lister groups.
The UI did not allow this but the API did. Only MCs should be able to change
Proof Listener and only MCs/BCs should be able to change Readers.

There was also a bug where if an MC edited another MCs profile it would remove
the other MC from the MC group. When updating a profile the previous code would
always remove all groups first and then add back the requested ones. But an MC
is not allowed to add an MC. There was code to try to work around this but it
only worked when an MC changed their own profile.

While trying to simplify things overall and fix all the above issues I decided
to make the update group process not delete all groups first. The code now
builds a list of groups you are allowed to change. It then adds or removes users
to these allowed groups only. This avoids the issue of needing special logic to
re-add groups you normally cannot change after they were all removed. It also
makes things more efficient as the group records in the database only change
when they really need to change instead of all the groups being removed and
re-added (deleteing and adding rows) when anything in the user profile changes
(or even if nothing changed in the update_profile call).