Fix reader URL link immediately after assigning section.
The reader array was being converted to a JSON string (stored in reader_id)
fairly early in the #assign_reader_btn function. This was then being passed
to set_reader_names which resulted in the ID in the URL having "[" and "]"
characters around the ID.

Instead we now keep the reader list as a JS array (reader_array). It is
converted to JSON string for the post but is accessed as a normal array
for the call to set_reader_names. This is more similar to how the
assign_section_array was already being handled.

Fixes #117.