Update cells in user table after profile is updated.
It can be a bit confusing when you just clicked the edit icon for a user
in the table, updated a value, clicked save and then the old value is still
displayed in the table.

This commit modifies the update_profile API so that it returns the fields
that were updated (it is possible that some fields sent to update_profile
were not updated because of lack of permissions). The edit_profile
function in profile_modal.js then uses this data to update the table.