Relax username validation for user profile updates.
Commit d406d197a7aa added more validation for the username field, including
the alpha_dash rule. However there are already existing user names that
violate this requirement in which case updating that profile would fail.

Note this only impacted full admin accounts as no one else (even MCs) can
attempt to edit a username. When you are not allowed to edit the username
no validation was taking place on the name field.