form_generators_authors: don't say empty auth_id
In what appears to be more fallout from STRICT_TRANS_TABLES, we can no
longer save an empty string auth_id into the database. Remove that
line of code and let the DB's autoincrement do its job. This appears
to happen only when an author that's already in the database is filled
in "manually" (ie, without clicking on the autocomplete proposal) in
the project template generator.