Reworks the pagination formatting
See issue #32. This commit changes the pagination at the bottom of the
page to better show how many pages are available. It doesn't include a
"Go to X" field for jumping to a specific page.

I've also tweaked the requested format slightly - if there are fewer
than ten pages, then we'll just show all of them.