Keep track of page numbers between refreshes
On a page of search results, we're tracking the current page number in
memory to know where we are. We also push that value into the browser
history under the query param `search_page`.

However, when a user browses away and comes back, we aren't passing that
query parameter through from the controller to the template that
renders the page, so we're just starting again from page 1 each time.

This commit modifes the frontend Javascript to look for the page number
in the query parameters rather than relying on the controller to inject
it.