Restore `lastest_releases` public URL endpoint
faa493f321b71879bcbbf0270cad0039fab93256 fixed the lastest -> latest
typo, but in doing do, inadvertently removed the public URL endpoint
for getting the latest releases. Because we don't know what code out
there had adapted to this typo, restore the endpoint by adding a shim
on top of the renamed method (we still want to continue using the
non-typo method name in the future for new code).