Remove PHP short tags
This commit removes the short tags (`<? ?>`) in favour of the long ones
(`<?php ?>`).

I was messing around with `ci-phpunit-test` to see if I could figure out
how to write unit tests for some of the PHP code, and I ran across some
warnings where the coverage tooling didn't understand short tags
properly.

As a bonus, this makes us more compliant with PSR-1.