Add ci-phpunit-test library
This adds ci-phpunit-test and phpunit so that we can write more in the
way of tests.

I'm using Composer for installing these, but I also did a manual install
version in another branch. (The result was pretty much the same, except
the vendor files end up checked in inside the
`application/tests/_ci_phpunit_test` directory.)