Render newlines in Project page description
For a while, newlines in Project descriptions weren't being rendered
properly in the catalog HTML, so people put manual `<br />` tags into
them.

Now, we want to render the newlines as they are in the admin screens,
but that could cause extra unintended whitespace for projects that
already have manual `<br />` tags in them, so we normalise all of the
different kinds of "newline" and strip some out if there are too many in
a row.