Update the CONTRIBUTING guide
It took me a fair bit of googling and playing around to get a grip on
how librivox-catalog fits together, so I thought I'd start writing some
of it down for the next person.

I haven't included any details on the structure of the code itself, but
I've included some links out to the CodeIgniter docs, which should go a
little way to helping people get started.